# services/Administracion/src/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MS Admin API",
    "description": "Microservicio de Administración - Configuración del negocio",
    "version": "1.0.0"
  },
  "paths": {
...
    "/mesas": {
      "get": {
        "tags": ["Mesas"],
        "summary": "Listar mesas de una sucursal",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "sucursal_id",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Lista de mesas" }
        }
      },
      "post": {
        "tags": ["Mesas"],
        "summary": "Crear una mesa",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearMesa"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Mesa creada" }
        }
      }
    },
    "/mesas/disponibles": {
      "get": {
        "tags": ["Mesas"],
        "summary": "Obtener mesas disponibles por capacidad",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "sucursal_id",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "capacidad",
            "in": "query",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Lista de mesas disponibles" }
        }
      }
    },
    "/mesas/{id}": {
      "get": {
        "tags": ["Mesas"],
        "summary": "Obtener una mesa por ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Mesa encontrada" }
        }
      },
      "put": {
        "tags": ["Mesas"],
        "summary": "Actualizar una mesa",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarMesa"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Mesa actualizada" }
        }
      },
      "delete": {
        "tags": ["Mesas"],
        "summary": "Eliminar una mesa",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Mesa eliminada" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CrearMesa": {
        "type": "object",
        "required": ["sucursal_id", "table_number", "capacity_min", "capacity_max"],
        "properties": {
          "sucursal_id": { "type": "string", "format": "uuid" },
          "table_number": { "type": "string" },
          "capacity_min": { "type": "integer", "minimum": 1 },
          "capacity_max": { "type": "integer", "minimum": 1 },
          "location": { "type": "string" }
        }
      },
      "ActualizarMesa": {
        "type": "object",
        "properties": {
          "table_number": { "type": "string" },
          "capacity_min": { "type": "integer", "minimum": 1 },
          "capacity_max": { "type": "integer", "minimum": 1 },
          "location": { "type": "string" },
          "is_active": { "type": "boolean" }
        }
      },
...
    }
  }
}